### 2019-08-21 #1673 @whedon
[PRE REVIEW]: EoN (Epidemics on Networks), software for simulation, analytic approximation, and analysis of epidemics on networks
**Submitting author:** @joelmiller (<a href="http://orcid.org/0000-0003-4426-0405">Joel Miller</a>)
**Repository:** <a href="https://github.com/springer-math/Mathematics-of-Epidemics-on-Networks" target ="_blank">https://github.com/springer-math/Mathematics-of-Epidemics-on-Networks</a>
**Version:** v1.0.8
