### 2019-08-21 #1673 @kyleniemeyer
@joelmiller this PR should fix the paper compilation issue: https://github.com/springer-math/Mathematics-of-Epidemics-on-Networks/pull/41

However, please also add more details to the affiliations, matching what you'd typically see in an article.
