### 2019-08-22 51c1a45
Fixed documentation to include hierarchy_pos
