+++ docs/functions/EoN.hierarchy_pos.rst
@@ -1,0 +1,6 @@
+EoN.hierarchy_pos
+=================
+
+.. currentmodule:: EoN
+
+.. autofunction:: hierarchy_pos
