### 2019-08-23 d347856
Edits of paper - removing typos and improving formatting.
