+++ docs/paper/paper.md
@@ -219,7 +219,8 @@ for cntr in range(10):
     print(cntr)
     print('doing Event-driven simulation')
     t, S, I, R = EoN.fast_nonMarkov_SIR(G, trans_time_fxn = trans_time_fxn,
-                        rec_time_fxn = rec_time_fxn_gamma, trans_time_args = (tau,))
+                                        rec_time_fxn = rec_time_fxn_gamma, 
+                                        trans_time_args = (tau,))
 
     #To reduce file size and make plotting faster, we'll just plot 1000
     #data points.  It's not really needed here, but this demonstrates
@@ -310,9 +311,9 @@ $\psi'(x)$ as well as $\phi_S(0)$, the probability
 an edge from a susceptible node connects to another susceptible node at time 0.
 By default, it assumes there are no recovered individuals at time $0$.
 
-If the population has a Poisson degree distristribution with mean $kave$ and the 
+If the population has a Poisson degree distristribution with mean ``kave`` and the 
 infection is introduced by randomly infecting a proportion ``rho`` of the population
-at time 0, then $\psi(x) = (1-\rho) e^{-rho(1-x)}$, 
+at time 0, then $\psi(x) = (1-\rho) e^{-\rho(1-x)}$, 
 $\psi'(x) = (1-\rho)\rho e^{-\rho(1-x)}$
 and $\phi_S(0) = 1-\rho$.  So
 
@@ -382,6 +383,7 @@ EoN provides a function ``Gillespie_simple_contagion`` which allows a user to
 specify the rules governing an arbitrary simple contagion.
 
 Examples are provided in the documentation, including
+
 - SEIR disease (there is an exposed state before becoming infectious)
 - SIRS disease (recovered individuals eventually become susceptible again)
 - SIRV disease (individuals may get vaccinated) 
@@ -427,7 +429,7 @@ We demonstrate this first with an SEIR example.  To demonstrate some of the
 flexibility we allow some individuals to have a higher rate of transitioning from 
 ``'E'`` to ``'I'`` and some partnerships to have a higher transmission rate. 
 This is done by adding weights to the contact network `G` which scale the 
-rates for those nodes or partners.  The documentation discusses other ways we 
+rates for those individuals or partnerships.  The documentation discusses other ways we 
 can allow for heterogeneity in transition rates.
 
 
@@ -453,8 +455,10 @@ G = nx.fast_gnp_random_graph(N, 5./(N-1))
 node_attribute_dict = {node: 0.5+random.random() for node in G.nodes()}
 edge_attribute_dict = {edge: 0.5+random.random() for edge in G.edges()}
 
-nx.set_node_attributes(G, values=node_attribute_dict, name='expose2infect_weight')
-nx.set_edge_attributes(G, values=edge_attribute_dict, name='transmission_weight')
+nx.set_node_attributes(G, values=node_attribute_dict, 
+                        name='expose2infect_weight')
+nx.set_edge_attributes(G, values=edge_attribute_dict, 
+                        name='transmission_weight')
 #
 #These individual and partnership attributes will be used to scale
 #the transition rates.  When we define `H` and `J`, we provide the name
@@ -470,7 +474,8 @@ H.add_edge('E', 'I', rate = 0.6, weight_label='expose2infect_weight')
 H.add_edge('I', 'R', rate = 0.1)
 
 J = nx.DiGraph()
-J.add_edge(('I', 'S'), ('I', 'E'), rate = 0.1, weight_label='transmission_weight')
+J.add_edge(('I', 'S'), ('I', 'E'), rate = 0.1, 
+            weight_label='transmission_weight')
 IC = defaultdict(lambda: 'S')
 for node in range(200):
     IC[node] = 'I'
@@ -497,8 +502,8 @@ This produces
 ![](SEIR.png)
     
[…]
