+++ docs/paper/paper.bib
@@ -52,7 +52,7 @@ series = {IAM},
 
 @techreport{hagberg2008exploring,
   title={Exploring network structure, dynamics, and function using NetworkX},
-  author={Hagberg, Aric and Swart, Pieter and S Chult, Daniel},
+  author={Hagberg, Aric and Swart, Pieter and Schult, Daniel},
   year={2008},
   institution={Los Alamos National Lab.(LANL), Los Alamos, NM (United States)}
 }
