### 2019-08-23 #1673 @stulacy
Unfortunately this is quite outside of my area. Either of the suggestions
made by @joelmiller look more appropriate.
