+++ docs/paper/paper.bib
@@ -108,6 +108,7 @@ doi = {10.1098/rsif.2011.0403}
 
 @MISC{stackoverflow:29586520,
     TITLE = {Can one get hierarchical graphs from networkx with python 3?},
+    Year = {2015},
     AUTHOR = {Joel C. Miller},
     HOWPUBLISHED = {Stack Overflow},
     NOTE = {URL:https://stackoverflow.com/a/29597209/2966723 (version: 2019-08-13)},
