### 2019-09-10 b541da1
Gillespie_simple_contagion allows directed graphs
