+++ docs/GettingStarted.rst
@@ -20,6 +20,19 @@ version available through pip (1.0.8).  The easiest way is
     pip install EoN
 
     
+The documentation provided here is for version 1.0.9rc1 which is on GitHub.  
+
+To **install EoN version 1.0.9rc1** You can clone or download the Github 
+version at https://github.com/springer-math/Mathematics-of-Epidemics-on-Networks. 
+
+Then just move into the main directory and run
+
+:: 
+
+    python setup.py install
+
+
+
 If you want to see changes from previous versions, please see 
 `Changes from v1.0 <Changes.html>`_.
 
