### 2019-09-11 96e0272
Adding setup.py to 1.0.9rc1
