### 2019-09-27 276124b
Minor edits for clarity of presentation.
