+++ docs/paper/paper.md
@@ -294,7 +294,7 @@ For ease of comparison with simulation, and consistency with existing literature
 the output of the model should be interpreted in terms of an expected number of individuals
 in each status, which requires that our values scale with ``N``.  So all of 
 the initial conditions have a factor of ``N``.  
-If we are interested in proportion,we could arbitrarily set ``N=1``, 
+If we are interested in proportion, we could arbitrarily set ``N=1``, 
 and then our solutions would give us the proportion of the population in each 
 status.
 
@@ -312,7 +312,7 @@ an edge from a susceptible node connects to another susceptible node at time 0.
 By default, it assumes there are no recovered individuals at time $0$.
 
 If the population has a Poisson degree distristribution with mean ``kave`` and the 
-infection is introduced by randomly infecting a proportion ``rho`` of the population
+infection is introduced by randomly infecting a proportion $\rho$ of the population
 at time 0, then $\psi(x) = (1-\rho) e^{-\rho(1-x)}$, 
 $\psi'(x) = (1-\rho)\rho e^{-\rho(1-x)}$
 and $\phi_S(0) = 1-\rho$.  So
@@ -391,7 +391,7 @@ Examples are provided in the documentation, including
 - Cooperative SIR diseases (infection with one disease helps spread the other)
 
 The implementation requires the user to separate out two distinct ways that 
-transitions occur.  To help demonstrate consider an "SEIR" epidemic, where 
+transitions occur.  To help demonstrate, consider an "SEIR" epidemic, where 
 individuals begin susceptible, but when they interact with infectious 
 individuals they may enter an exposed state.  They remain in that exposed 
 state for some period of time before transitioning into the infectious state.
@@ -399,25 +399,27 @@ They remain infectious and eventually transition into the recovered state.
 
 We can identify two broad types of transitions:
 
-- Sometimes individuals change status without influence from any other individual.
-  For example, an infected individual may recover, or an exposed individual may 
-  move into the infectious class.  These transitions between statuses can be 
-  represented by a directed graph ``H`` where the nodes are not the original 
-  individuals of the contact network ``G``, but rather the potential statuses 
-  individuals can take.  The edges represent transitions that can occur, and 
-  we weight the edges by the rate.  In the SEIR case we would need the graph 
-  ``H`` to have edges ``'E'``->``'I'`` and ``'I'``->``'R'``.  The
-  edges would be weighted by the transition rates.  Note ``H`` need not have 
-  a node ``'S'`` because susceptible nodes do not change status on their own.
+- **Spontaneous Transitions**  Sometimes individuals change status without 
+  influence from any other individual.  For example, an infected individual 
+  may recover, or an exposed individual may move into the infectious class.  
+  These transitions between statuses can be represented by a directed graph 
+  ``H`` where the nodes are not the original individuals of the contact 
+  network ``G``, but rather the potential statuses individuals can take.  The 
+  edges represent transitions that can occur, and we weight the edges by the 
+  rate.  In the SEIR case we would need the graph ``H`` to have edges 
+  ``'E'``->``'I'`` and ``'I'``->``'R'``.  The edges would be weighted by the 
+  transition rates.  Note ``H`` need not have a node ``'S'`` because 
+  susceptible nodes do not change status on their own.
 
-- Sometimes individuals change status due to the influence of a single partner.
[…]
