+++ docs/GettingStarted.rst
@@ -20,7 +20,8 @@ version available through pip (1.0.8).  The easiest way is
     pip install EoN
 
     
-The documentation provided here is for version 1.0.9rc1 which is on GitHub.  
+The documentation provided here is for version 1.0.9rc1 which is on GitHub (see
+`Changes from v1.0 <Changes.html>`_ for any relevant changes).
 
 To **install EoN version 1.0.9rc1** You can clone or download the Github 
 version at https://github.com/springer-math/Mathematics-of-Epidemics-on-Networks. 
