### 2019-11-25 #1731 @lpantano
Hi @hagberg, did you end up opening the issues to ask the authors to change issues you saw during the review? If yes, can you link them here? Thanks!
