### 2019-11-26 springer-math/Mathematics-of-Epidemics-on-Networks#43 @hagberg
Functional testing
Are there methods (preferably automated) to test the functionality of the code?
It would be good to be able to verify at minimum that the basic installation is functioning.
Related issue at https://github.com/openjournals/joss-reviews/issues/1731
