### 2019-11-26 springer-math/Mathematics-of-Epidemics-on-Networks#44 @hagberg
Adding related work and refs to paper
The draft paper is missing a description of and references to related software packages.  See issue at https://github.com/openjournals/joss-reviews/issues/1731
