### 2019-12-02 springer-math/Mathematics-of-Epidemics-on-Networks#43 @joelmiller
I have merged the functional testing created by @tinghf .  I believe this will address this issue.

The tests are now under EoN/tests, and there is a readme file there created by @tinghf
