### 2019-12-02 c8029e1
Small change to hierarchy_pos to remove irrelevant print statement.  Updates to paper and documentation.
