+++ docs/paper/paper.md
@@ -39,8 +39,8 @@ book *Mathematics of Epidemics on Networks* [@kiss:EoN], and now consists of ove
 
 EoN provides a set of tools for
 
-- Susceptible-Infected-Susceptible (SIS) and Susceptible-Infected-Recovered (SIR)
-disease
+- Susceptible-Infected-Susceptible (SIS) and Susceptible-Infected-Recovered 
+(SIR) disease
   - Stochastic simulation of disease spread in networkx graphs
     - continuous time Markovian
     - continuous time nonMarkovian
@@ -53,29 +53,32 @@ disease
 - Visualization and analysis of stochastic simulations
 
 These algorithms are built on the networkx package [@hagberg2008exploring].
-Here we provide brief descriptions with examples of a few of these tools.
 EoN's documentation is maintained at 
 https://epidemicsonnetworks.readthedocs.io/en/latest/ 
 including numerous examples at 
-https://epidemicsonnetworks.readthedocs.io/en/latest/Examples.html.
+https://epidemicsonnetworks.readthedocs.io/en/latest/Examples.html.  In this 
+paper we provide brief descriptions with examples of a few of EoN's tools.
 
-We model spreading processes on a contact network, and it is often useful for
-mathematicians and physicists to formally think of individuals as nodes with 
-their potentially infectious partnerships as edges.  However, for those who 
-come from other backgrounds this abstraction may be less familiar.  Therefore, 
-we will describe a contact network along which an infections process spreads 
-as consisting of "individuals" and "partnerships" rather than "nodes" and 
-"edges".  This has an additional benefit because in the simple contagion code, 
-we need to define some other networks whose nodes represent possible statuses
-and whose edges represent transitions that can occur.  Referring to "individuals" 
-and "partnerships" when discussing the process spreading on the contact network 
-makes it easier to avoid confusion between the different networks.
+We model spreading processes on a contact network.  In this context, many
+mathematicians and physicists are accustomed to thinking of individuals as 
+nodes with their potentially infectious partnerships as edges.  However, for 
+those who come from other backgrounds this abstraction may be less familiar.  
+Therefore, we will describe a contact network along which an infections 
+process spreads as consisting of "individuals" and "partnerships" rather than 
+"nodes" and "edges".  This has an additional benefit because in the simple 
+contagion algorithm (described later), we need to define some other networks 
+whose nodes represent possible statuses and whose edges represent transitions 
+that can occur.  Referring to "individuals" and "partnerships" when discussing
+the process spreading on the contact network makes it easier to avoid 
+confusion between the different networks.
 
 We start with a description of the tools for studying SIS and SIR disease 
 through stochastic simulation and differential equations models.  Then we 
 describe the simple and complex contagions, including examples showing how
 the simple contagion can be used to capture a range of standard disease
-models.  Finally we demonstrate the visualization software.
+models.  Finally we demonstrate the visualization tools.  The examples shown
+are intended to demonstrate the ability of the tools.  The documentation gives
+more detail about how to use them.
 
 ## SIR and SIS disease
 
@@ -93,7 +96,7 @@ It is possible for transition rates to depend on intrinsic properties of
 individuals and of partnerships.
 
 The continuous-time stochastic simulations have two different implementations: a 
-Gillespie implementation [@gillespie1977exact, @doob1945markoff] and an Event-driven
+Gillespie implementation [@gillespie1977exact,@doob1945markoff] and an Event-driven
 implementation.  Both approaches are efficient.  They have similar speed if the 
 dynamics are Markovian (depending on the network and disease parameters either
[…]
