+++ docs/conf.py
@@ -61,7 +61,7 @@ author = u'Joel Miller, Istvan Kiss, Peter Simon'
 # The short X.Y version.
 version = u'1.0.9'
 # The full version, including alpha/beta/rc tags.
-release = u'1.0.9rc1' 
+release = u'1.0.9rc2' 
 
 # The language for content autogenerated by Sphinx. Refer to documentation
 # for a list of supported languages.
