+++ docs/paper/paper.bib
@@ -31,6 +31,17 @@
   publisher={JSTOR}
 }
 
+@article{jenness:EpiModel,
+title = {{EpiModel}: An {R} Package for Mathematical 
+Modeling of Infectious Disease over Networks},
+author = {S. M. Jenness and  S. M. Goodreau and M. Morris},
+journal = {Journal of Statistical Software}
+year = {2018},
+volume = {84},
+number = {8},
+pages = {1--47}
+}
+
 @book{kiss:EoN,
   title={Mathematics of epidemics on networks: from exact to approximate models},
   author={Kiss, Istvan Z and Miller, Joel C and Simon, P{\'e}ter L},
@@ -106,6 +117,16 @@ doi = {10.1098/rsif.2011.0403}
   publisher={Oxford University Press}
 }
 
+@article{peixoto_graph-tool_2014,
+         title = {The graph-tool python library},
+         url = {http://figshare.com/articles/graph_tool/1164194},
+         doi = {10.6084/m9.figshare.1164194},
+         urldate = {2014-09-10},
+         journal = {figshare},
+         author = {Peixoto, Tiago P.},
+         year = {2014},
+         keywords = {all, complex networks, graph, network, other}}
+         
 @MISC{stackoverflow:29586520,
     TITLE = {Can one get hierarchical graphs from networkx with python 3?},
     Year = {2015},
