+++ docs/paper/paper.bib
@@ -35,7 +35,7 @@
 title = {{EpiModel}: An {R} Package for Mathematical 
 Modeling of Infectious Disease over Networks},
 author = {S. M. Jenness and  S. M. Goodreau and M. Morris},
-journal = {Journal of Statistical Software}
+journal = {Journal of Statistical Software},
 year = {2018},
 volume = {84},
 number = {8},
