### 2019-12-03 #1731 @joelmiller
Hi @lpantano - Do you know if there is any way to embed a video into the pdf file?  One of the final things I want to show is a tool to build an animation, but at present I can't find any way to embed it.
