### 2019-12-04 77f7124
fixing setup.py to (hopefully) fix readthedocs error.
