### 2019-12-04 #1731 @joelmiller
@hagberg

I've added the functional testing (created by @tinghf )  into the repository under EoN/tests

I've also added a discussion of related packages into the paper.

Let me know of anything else you'd like improved.
