### 2019-12-04 748c074
update readthedocs.yml to no longer specify the python version (since networkx now wants a more recent version).
