+++ readthedocs.yml
@@ -4,5 +4,5 @@ build:
   image: latest
 
 python:
-  version: 3.5
+#  version: 3.5
   setup_py_install: true
