### 2019-12-04 4d638d3
another try at fixing the readthedocs.yml file.
