### 2019-12-05 388592a
reverting setup.py to state 1.0.9rc2 (had changed to 1.0.9) now that it's clear that the issue with readthedocs was from networkx's setup.py, not ours.
