### 2019-12-05 springer-math/Mathematics-of-Epidemics-on-Networks#44 @hagberg
Fixed in
https://github.com/springer-math/Mathematics-of-Epidemics-on-Networks/commit/4ef822bc94ade0f74032d69eb7469266501a9109
