### 2019-12-05 #1731 @hagberg
Looks good.  Issues are closed.
