### 2019-12-06 d7f67c7
small changes in paper
