### 2019-12-06 61f448a
remove animation from paper.md (since it does not embed)
