### 2019-12-06 #1731 @joelmiller
@lpantano @danielskatz

Okay - I've removed the animation from the paper.md file.  Should be good to go now.
