### 2019-12-06 #1731 @lpantano
@joelmiller, it seems there are missing DOIs, can you double check this? Thanks!
