### 2019-12-09 5a43339
Small edits to paper and add DOIs to bibliography
