+++ docs/paper/paper.md
@@ -57,12 +57,15 @@ EoN's documentation is maintained at
 https://epidemicsonnetworks.readthedocs.io/en/latest/ 
 including numerous examples at 
 https://epidemicsonnetworks.readthedocs.io/en/latest/Examples.html.  In this 
-paper we provide brief descriptions with examples of a few of EoN's tools.
+paper we provide brief descriptions with examples of a few of EoN's tools.  
+The examples shown are intended to demonstrate the ability of the tools.  The 
+online documentation gives more detail about how to use them.
 
 We model spreading processes on a contact network.  In this context, many
 mathematicians and physicists are accustomed to thinking of individuals as 
 nodes with their potentially infectious partnerships as edges.  However, for 
 those who come from other backgrounds this abstraction may be less familiar.  
+
 Therefore, we will describe a contact network along which an infections 
 process spreads as consisting of "individuals" and "partnerships" rather than 
 "nodes" and "edges".  This has an additional benefit because in the simple 
@@ -72,27 +75,26 @@ that can occur.  Referring to "individuals" and "partnerships" when discussing
 the process spreading on the contact network makes it easier to avoid 
 confusion between the different networks.
 
-We start with a description of the tools for studying SIS and SIR disease 
+We start this paper by describing the tools for studying SIS and SIR disease 
 through stochastic simulation and differential equations models.  Then we 
-describe the simple and complex contagions, including examples showing how
+discuss the simple and complex contagions, including examples showing how
 the simple contagion can be used to capture a range of standard disease
-models.  Finally we demonstrate the visualization tools.  The examples shown
-are intended to demonstrate the ability of the tools.  The documentation gives
-more detail about how to use them.
+models.  Finally we demonstrate the visualization tools.
 
 ## SIR and SIS disease
 
 ### Stochastic simulation
 The stochastic SIR and SIS simulation tools allow the user to investigate  
 standard SIS and SIR dynamics (SEIR/SIRS and other processes are addressed 
-below):
+within the simple contagion model):
 
 - Markovian SIS and SIR simulations  (``fast_SIS``, ``Gillespie_SIS``, ``fast_SIR``, and ``Gillespie_SIR``).
 - non-Markovian SIS and SIR simulations (``fast_nonMarkovian_SIS`` and ``fast_nonMarkovian_SIR``).
 - discrete time SIS and SIR simulations where infections last a single time step 
   (``basic_discrete_SIS``, ``basic_discrete_SIR``, and ``discrete_SIR``).
 
-It is possible for transition rates to depend on intrinsic properties of 
+For both Markovian and non-Markovian methods it is possible for the transition 
+rates to depend on intrinsic properties of 
 individuals and of partnerships.
 
 The continuous-time stochastic simulations have two different implementations: a 
@@ -102,8 +104,9 @@ dynamics are Markovian (depending on the network and disease parameters either
 may be faster than the other), but the event-driven implementation can also handle 
 non-Markovian dynamics.  In earlier versions, the event-driven simulations were 
 consistently faster than the Gillespie simulations, and thus they are named 
-`fast_SIR` and `fast_SIS`.  The Gillespie simulations were sped up using ideas
-from [@holme2014model] and [@cota2017optimized].
+`fast_SIR` and `fast_SIS`.  The Gillespie simulations have since reached
+comparable speed using up using ideas from [@holme2014model] and 
+[@cota2017optimized].
 
 The algorithms can typically handle an SIR epidemic spreading on 
 hundreds of thousands of individuals within a few seconds on a laptop.  The SIS 
@@ -135,6 +138,9 @@ gamma = 1.0 #recovery rate
 
[…]
