+++ docs/paper/paper.bib
@@ -7,6 +7,7 @@
   number={1},
   pages={449--456},
   year={2007},
+  doi = {10.1016/j.physa.2006.06.018},
   publisher={Elsevier}
 }
 
@@ -17,6 +18,7 @@
   volume={219},
   pages={303--312},
   year={2017},
+  doi = {10.1016/j.cpc.2017.06.007},
   publisher={Elsevier}
 }
 
@@ -28,6 +30,7 @@
   number={3},
   pages={455--473},
   year={1945},
+  doi = {10.1090/S0002-9947-1945-0013857-4},
   publisher={JSTOR}
 }
 
@@ -39,7 +42,8 @@ journal = {Journal of Statistical Software},
 year = {2018},
 volume = {84},
 number = {8},
-pages = {1--47}
+pages = {1--47},
+doi = {10.18637/jss.v084.i08}
 }
 
 @book{kiss:EoN,
@@ -47,7 +51,8 @@ pages = {1--47}
   author={Kiss, Istvan Z and Miller, Joel C and Simon, P{\'e}ter L},
 publisher = {Springer},
 series = {IAM},
-  year={2017}
+  year={2017},
+  doi = {10.1007/978-3-319-50806-1}
 }
 @article{gillespie1977exact,
 	Author = {Gillespie, D. T.},
@@ -58,14 +63,19 @@ series = {IAM},
 	Pages = {2340--2361},
 	Title = {Exact Stochastic Simulation of Coupled Chemical Reactions},
 	Volume = {81},
-	Year = {1977}}
+	Year = {1977},
+	doi = {10.1021/j100540a008}
+	}
 
 
-@techreport{hagberg2008exploring,
+@inproceedings{hagberg2008exploring,
   title={Exploring network structure, dynamics, and function using NetworkX},
-  author={Hagberg, Aric and Swart, Pieter and Schult, Daniel},
+  booktitle =   {Proceedings of the 7th Python in Science Conference},
+  author={Hagberg, Aric A.  and Schult, Daniel A. and Swart, Pieter J.},
+  pages={11-15},
   year={2008},
-  institution={Los Alamos National Lab.(LANL), Los Alamos, NM (United States)}
+  editor =    {Ga\"el Varoquaux and Travis Vaught and Jarrod Millman},
+  doi={10.25080/issn.2575-9752}
 }
 
 @article{holme2014model,
@@ -74,12 +84,14 @@ series = {IAM},
   journal={arXiv preprint arXiv:1403.1011},
   year={2014}
 }
+
 @article{miller:contagion,
   title={Complex contagions and hybrid phase transitions},
   author={Miller, Joel C},
   journal={Journal of Complex Networks},
   pages={cnv021},
   year={2016},
+  doi={10.1093/comnet/cnv021},
   publisher={Oxford University Press}
 }
 
@@ -114,6 +126,7 @@ doi = {10.1098/rsif.2011.0403}
   number={11},
   pages={1852--1861},
   year={2018},
+  doi = {10.1093/bioinformatics/bty921},
   publisher={Oxford University Press}
 }
 
@@ -132,9 +145,9 @@ doi = {10.1098/rsif.2011.0403}
     Year = {2015},
     AUTHOR = {Joel C. Miller},
     HOWPUBLISHED = {Stack Overflow},
-    NOTE = {URL:https://stackoverflow.com/a/29597209/2966723 (version: 2019-08-13)},
-    EPRINT = {https://stackoverflow.com/a/29597209/2966723},
-    URL = {https://stackoverflow.com/a/29597209/2966723}
+    NOTE = {URL:https://stackoverflow.com/revisions/29597209/14},
[…]
