+++ docs/paper/paper.md
@@ -57,7 +57,7 @@ EoN's documentation is maintained at
 https://epidemicsonnetworks.readthedocs.io/en/latest/ 
 including numerous examples at 
 https://epidemicsonnetworks.readthedocs.io/en/latest/Examples.html.  In this 
-paper we provide brief descriptions with examples of a few of EoN's tools.  
+paper we provide brief descriptions with examples of a few of EoN's tools.
 The examples shown are intended to demonstrate the ability of the tools.  The 
 online documentation gives more detail about how to use them.
 
@@ -84,7 +84,7 @@ models.  Finally we demonstrate the visualization tools.
 ## SIR and SIS disease
 
 ### Stochastic simulation
-The stochastic SIR and SIS simulation tools allow the user to investigate  
+The stochastic SIR and SIS simulation tools allow the user to investigate
 standard SIS and SIR dynamics (SEIR/SIRS and other processes are addressed 
 within the simple contagion model):
 
@@ -105,12 +105,11 @@ may be faster than the other), but the event-driven implementation can also hand
 non-Markovian dynamics.  In earlier versions, the event-driven simulations were 
 consistently faster than the Gillespie simulations, and thus they are named 
 `fast_SIR` and `fast_SIS`.  The Gillespie simulations have since reached
-comparable speed using up using ideas from [@holme2014model] and 
-[@cota2017optimized].
+comparable speed using ideas from [@holme2014model] and [@cota2017optimized].
 
 The algorithms can typically handle an SIR epidemic spreading on 
-hundreds of thousands of individuals within a few seconds on a laptop.  The SIS 
-versions are slower because the number of events that can happen is much
+hundreds of thousands of individuals in well under a minute on a laptop.  The 
+SIS versions are slower because the number of events that can happen is much
 larger in an SIS simulation.
 
 #### Examples
@@ -254,14 +253,14 @@ This produces a (stochastic) figure like
 EoN also provides a set of tools to numerically solve approximately 20 differential equations
 models for SIS or SIR disease spread in networks.  The various models use different
 information about the network to make deterministic predictions about
-the fraction infected at different times.  These use the Scipy integration tools.
+the number infected at different times.  These use the Scipy integration tools.
 The derivations of the models and explanations of their simplifying assumptions
 are described in [@kiss:EoN].
 
 Depending on the model, we need different information about the network structure.
 The algorithms allow us to provide the information as inputs.  However, there
 is also a version of each model which takes a network as an input 
-instead and then measure the network properties.
+instead and then measures the network properties.
 
 #### Examples
 
@@ -312,7 +311,7 @@ Our second example uses an Edge-based compartmental model for an SIR disease
 (section 6.5.2 of [@kiss:EoN]).
 This model incorporates information about the degree distribution (i.e., how the
 number of partners is distributed), but assumes that the partnerships are selected
-as randomly as possible given this distribution.  This requires we define the
+as randomly as possible given this distribution.  The model requires we define a
 "generating function" $\psi(x)$ which is equal to the sum 
 $\sum_{k=0}^\infty S_k(0) x^k$ where
 $S_k(0)$ is the proportion of all individuals in the population who both have $k$
@@ -323,8 +322,8 @@ By default, it assumes there are no recovered individuals at time $0$.
 
 If the population has a Poisson degree distristribution with mean ``kave`` and the 
 infection is introduced by randomly infecting a proportion $\rho$ of the population
[…]
