### 2019-12-09 cc1d322
Fine tuning of paper
