### 2019-12-09 86e8381
Improve paper's cooperative diseases figure and example.
