+++ docs/paper/paper.bib
@@ -22,6 +22,17 @@
   publisher={Elsevier}
 }
 
+@article{cui2017mutually,
+  title={Mutually cooperative epidemics on power-law networks},
+  author={Cui, Peng-Bi and Colaiori, Francesca and Castellano, Claudio and others},
+  journal={Physical Review E},
+  volume={96},
+  number={2},
+  pages={022301},
+  year={2017},
+  doi = {10.1103/PhysRevE.96.022301},
+  publisher={APS}
+}
 @article{doob1945markoff,
   title={Markoff chains--denumerable case},
   author={Doob, Joseph L},
@@ -46,14 +57,7 @@ pages = {1--47},
 doi = {10.18637/jss.v084.i08}
 }
 
-@book{kiss:EoN,
-  title={Mathematics of epidemics on networks: from exact to approximate models},
-  author={Kiss, Istvan Z and Miller, Joel C and Simon, P{\'e}ter L},
-publisher = {Springer},
-series = {IAM},
-  year={2017},
-  doi = {10.1007/978-3-319-50806-1}
-}
+
 @article{gillespie1977exact,
 	Author = {Gillespie, D. T.},
 	Date-Added = {2014-02-08 09:17:11 +0000},
@@ -68,6 +72,17 @@ series = {IAM},
 	}
 
 
+@article{grassberger2016phase,
+  title={Phase transitions in cooperative coinfections: Simulation results for networks and lattices},
+  author={Grassberger, Peter and Chen, Li and Ghanbarnejad, Fakhteh and Cai, Weiran},
+  journal={Physical Review E},
+  volume={93},
+  number={4},
+  pages={042316},
+  year={2016},
+  doi = {10.1103/PhysRevE.93.042316},
+  publisher={APS}
+}
 @inproceedings{hagberg2008exploring,
   title={Exploring network structure, dynamics, and function using NetworkX},
   booktitle =   {Proceedings of the 7th Python in Science Conference},
@@ -83,6 +98,39 @@ series = {IAM},
   author={Holme, Petter},
   journal={arXiv preprint arXiv:1403.1011},
   year={2014}
+}
+
+@book{kiss:EoN,
+  title={Mathematics of epidemics on networks: from exact to approximate models},
+  author={Kiss, Istvan Z and Miller, Joel C and Simon, P{\'e}ter L},
+publisher = {Springer},
+series = {IAM},
+  year={2017},
+  doi = {10.1007/978-3-319-50806-1}
+}
+@article{liu2018interactive,
+  title={Interactive social contagions and co-infections on complex networks},
+  author={Liu, Quan-Hui and Zhong, Lin-Feng and Wang, Wei and Zhou, Tao and Eugene Stanley, H},
+  journal={Chaos: An Interdisciplinary Journal of Nonlinear Science},
+  volume={28},
+  number={1},
+  pages={013120},
+  year={2018},
+  doi={10.1063/1.5010002},
+  publisher={AIP Publishing}
+}
+
+
+@article{miller:compete,
+  title={Cocirculation of infectious diseases on networks},
+  author={Miller, Joel C.},
+  journal={Physical Review E},
+  volume={87},
+  number={6},
+  pages={060801},
+  year={2013},
+  doi={10.1103/PhysRevE.87.060801},
+  publisher={APS}
 }
 
 @article{miller:contagion,
