### 2019-12-10 f1c7684
improving comments in code blocks
