### 2019-12-10 #1731 @joelmiller
@lpantano - I think this is good to go now.  I've updated the references and gone through the paper a couple times.
