### 2019-12-10 #1731 @joelmiller
That's actually a bit problematic.  The software was developed initially alongside a textbook I was writing, so it sits within https://github.com/springer-math.  I've sent a request to give Zenodo the necessary access.  But the editor who set up that site has left, and I'm not even sure that anyone will get the message.

Thoughts?  The obvious idea would be to create a new repository.  But I prefer to move things around as little as possible.
