### 2019-12-10 #1731 @joelmiller
@lpantano

That's actually a bit problematic. The software was developed initially alongside a textbook I was writing, so it sits within https://github.com/springer-math (which I don't own). I've sent a request to give Zenodo the necessary access. But the editor who set up that site has left the company, and I'm not even sure that anyone will get the message.  It looks like it hasn't been used since he left (over a year ago).

Thoughts? The obvious idea would be to create a new repository. But I prefer to move things around as little as possible.
