### 2019-12-10 #1731 @danielskatz
You can simply tar the repo and manually deposit it in Zenodo (or figshare, or any other archival repository that will accept it and provide a DOI)

This is what the GitHub - Zenodo link automates, but you don't have to use the automatic mechanism
