+++ docs/conf.py
@@ -59,9 +59,9 @@ author = u'Joel Miller, Istvan Kiss, Peter Simon'
 # built documents.
 #
 # The short X.Y version.
-version = u'1.0.9'
+version = u'1.1'
 # The full version, including alpha/beta/rc tags.
-release = u'1.0.9rc2' 
+release = u'1.1' 
 
 # The language for content autogenerated by Sphinx. Refer to documentation
 # for a list of supported languages.
