+++ EoN/__init__.py
@@ -46,7 +46,7 @@ We start with a few useful auxiliary functions
 '''
 
 __author__ = "Joel C. Miller, with tests written by Tony Ting"
-__version__ = "1.0.9rc2"
+__version__ = "1.1"
 def __citation__():
     print("A proper citation should be coming soon\n" + \
               "Until then consider citing the github page\n" + \
