+++ docs/GettingStarted.rst
@@ -4,7 +4,7 @@ Getting Started
 
 Installation
 ------------
-You can **install EoN version 1.0.8** with pip
+You can **install EoN version 1.1** with pip
 
 ::
 
@@ -12,7 +12,7 @@ You can **install EoN version 1.0.8** with pip
 
 
 If you have installed a previous version and want to reinstall the most recent
-version available through pip (1.0.8).  The easiest way is 
+version available through pip (1.1).  The easiest way is 
 
 ::
 
@@ -20,11 +20,10 @@ version available through pip (1.0.8).  The easiest way is
     pip install EoN
 
     
-The documentation provided here is for version 1.0.9rc2 which is on GitHub (see
-`Changes from v1.0 <Changes.html>`_ for any relevant changes).
 
-To **install EoN version 1.0.9rc2** You can clone or download the Github 
-version at https://github.com/springer-math/Mathematics-of-Epidemics-on-Networks. 
+To **install EoN a later (in development) version** You can clone or download 
+the Github version at 
+https://github.com/springer-math/Mathematics-of-Epidemics-on-Networks
 
 Then just move into the main directory and run
 
@@ -33,6 +32,7 @@ Then just move into the main directory and run
     python setup.py install
 
 
+The documentation provided here is for version 1.1.
 
 If you want to see changes from previous versions, please see 
 `Changes from v1.0 <Changes.html>`_.
