+++ docs/Changes.rst
@@ -1,12 +1,9 @@
 Changes from v 1.0
 ==================
 
-New in v 1.0.9rc2
+New in v 1.1
 -----------------
 - ``Hierarchy_Pos`` has an extraneous print statement removed.
-
-New in v 1.0.9rc1
------------------
 
 - ``Gillespie_simple_contagion`` should now accept a directed graph ``G``.
 
