+++ README.md
@@ -17,6 +17,8 @@ We released v1.0 on 9 Aug 2018
 To see changes since v1.0, please look at our 
 [list of changes](file:///Users/jomiller/Dropbox/work/SpringerGithub/docs/_build/html/Changes.html)
 
+We released v1.1 on 13 Dec 2019.  This is the version described in our paper.
+
 ## Corrections
 
 For corrections or to report an error to the content in the published book, see the file [errata.md](errata.md) and follow instructions there.
