### 2019-12-12 f113841
small documentation update (where to cite)
