### 2019-12-12 #1731 @joelmiller
@danielskatz @lpantano

2 things;

- I've updated the version to 1.1 (from 1.0.8)
- I've reserved the DOI 10.5281/zenodo.3572756.  I'd like to know the final paper reference before I finalize the submission to Zenodo.  Will that be feasible?  Otherwise I'll just go ahead and submit it.
