### 2019-12-13 b4d4a6a
Add citation info with DOI of JOSS publication.
