### 2019-12-16 #1731 @joelmiller
@lpantano, I've made those changes.  The only remaining inconsistency is that after updating changes I decided to rename this as v1.1.  I don't see the version appearing anywhere in the proof so it's probably not an issue, but I do see that in the metadata in the first post on this thread it gives the version number.  So if it matters we should update, but if not, I think we're done.
