### 2019-12-16 9befe4e
update citation info in __init__.py
