### 2019-12-17 #1731 @lpantano
@openjournals/joss-eics, I think this is good to get an eye from you.
