### 2019-12-18 #1731 @arfon
@lpantano @joelmiller - this paper is _very_ long for JOSS (24 pages). From a quick review, there look to been a substantial amount of documentation (e.g. code examples) that would be more appropriate in supporting documentation.

What do you think about moving this content into the docs?
