### 2019-12-18 #1731 @labarba
In fact, the [JOSS documentation](https://joss.readthedocs.io/en/latest/submitting.html) stipulates that papers should be _short_: "The paper should be between 250-1000 words." And also,

> a “full length” paper is not permitted, and software documentation such as API (Application Programming Interface) functionality should not be in the paper and instead should be outlined in the software documentation.

TL;DR: JOSS does not publish long-form papers.
