### 2019-12-19 41f5cdc
Dramatically shorten paper by removing examples.
