+++ docs/paper/paper.md
@@ -61,26 +61,7 @@ paper we provide brief descriptions with examples of a few of EoN's tools.
 The examples shown are intended to demonstrate the ability of the tools.  The 
 online documentation gives more detail about how to use them.
 
-We model spreading processes on a contact network.  In this context, many
-mathematicians and physicists are accustomed to thinking of individuals as 
-nodes with their potentially infectious partnerships as edges.  However, for 
-those who come from other backgrounds this abstraction may be less familiar.  
-
-Therefore, we will describe a contact network along which an infections 
-process spreads as consisting of "individuals" and "partnerships" rather than 
-"nodes" and "edges".  This has an additional benefit because in the simple 
-contagion algorithm (described later), we need to define some other networks 
-whose nodes represent possible statuses and whose edges represent transitions 
-that can occur.  Referring to "individuals" and "partnerships" when discussing
-the process spreading on the contact network makes it easier to avoid 
-confusion between the different networks.
-
-We start this paper by describing the tools for studying SIS and SIR disease 
-through stochastic simulation and differential equations models.  Then we 
-discuss the simple and complex contagions, including examples showing how
-the simple contagion can be used to capture a range of standard disease
-models.  Finally we demonstrate the visualization tools.
-
+We model spreading processes on a contact network.  
 ## SIR and SIS disease
 
 ### Stochastic simulation
@@ -104,149 +85,13 @@ dynamics are Markovian (depending on the network and disease parameters either
 may be faster than the other), but the event-driven implementation can also handle 
 non-Markovian dynamics.  In earlier versions, the event-driven simulations were 
 consistently faster than the Gillespie simulations, and thus they are named 
-`fast_SIR` and `fast_SIS`.  The Gillespie simulations have since reached
-comparable speed using ideas from [@holme2014model] and [@cota2017optimized].
+`fast_SIR` and `fast_SIS`.  The Gillespie simulations have since been optimized
+using ideas from [@holme2014model] and [@cota2017optimized].
 
 The algorithms can typically handle an SIR epidemic spreading on 
 hundreds of thousands of individuals in well under a minute on a laptop.  The 
 SIS versions are slower because the number of events that happen is often much
 larger in an SIS simulation.
-
-#### Examples
-
-To demonstrate these, we begin with SIR simulations on an Erdős-Rényi network
-having a million individuals (in an Erdős-Rényi network each individual has 
-identical probability of independently partnering with any other individual in 
-the population).
-
-
-```python
-import networkx as nx
-import EoN
-import matplotlib.pyplot as plt
-
-
-N = 10**6  #number of individuals
-kave = 5    #expected number of partners
-print('generating graph G with {} nodes'.format(N))
-G = nx.fast_gnp_random_graph(N, kave/(N-1)) #Erdős-Rényi graph
-    
-rho = 0.005 #initial fraction infected
-tau = 0.3   #transmission rate
-gamma = 1.0 #recovery rate
-
-print('doing event-based simulation')
-t1, S1, I1, R1 = EoN.fast_SIR(G, tau, gamma, rho=rho)
-#instead of rho, we could specify a list of nodes as ``initial_infecteds``, or
-#specify neither and a single random node would be chosen as the index case.
-
-print('doing Gillespie simulation')
[…]
