### 2019-12-19 3a9a8be
More shortening of paper
