+++ docs/paper/paper.md
@@ -31,8 +31,8 @@ bibliography: paper.bib
 
 # Summary
 
-EoN (EpidemicsOnNetworks) is a pure-python package designed to assist studies of
-infectious processes spreading through networks.  It originally rose out of the 
+EoN (EpidemicsOnNetworks) is a pure-python package designed to study
+infectious processes spreading in networks.  It rose out of the 
 book *Mathematics of Epidemics on Networks* [@kiss:EoN], and now consists of over 
 100 user-functions.
 
@@ -52,20 +52,18 @@ EoN provides a set of tools for
 - Stochastic simulation of a wide range of Simple and Complex contagions
 - Visualization and analysis of stochastic simulations
 
-These algorithms are built on the networkx package [@hagberg2008exploring].
-EoN's documentation is maintained at 
+EoN relies on the networkx package [@hagberg2008exploring].
+Its documentation is maintained at 
 https://epidemicsonnetworks.readthedocs.io/en/latest/ 
 including numerous examples at 
 https://epidemicsonnetworks.readthedocs.io/en/latest/Examples.html.  In this 
-paper we provide brief descriptions with examples of a few of EoN's tools.
-The examples shown are intended to demonstrate the ability of the tools.  The 
-online documentation gives more detail about how to use them.
+paper we provide brief descriptions of a few of EoN's main tools.  The 
+online documentation gives more detail about how to use them, including examples.
 
-We model spreading processes on a contact network.  
 ## SIR and SIS disease
 
 ### Stochastic simulation
-The stochastic SIR and SIS simulation tools allow the user to investigate
+The main stochastic simulation tools allow the user to investigate
 standard SIS and SIR dynamics (SEIR/SIRS and other processes are addressed 
 within the simple contagion model):
 
@@ -75,60 +73,57 @@ within the simple contagion model):
   (``basic_discrete_SIS``, ``basic_discrete_SIR``, and ``discrete_SIR``).
 
 For both Markovian and non-Markovian methods it is possible for the transition 
-rates to depend on intrinsic properties of 
-individuals and of partnerships.
+rates to depend on individual or partnership properties.
 
 The continuous-time stochastic simulations have two different implementations: a 
 Gillespie implementation [@gillespie1977exact; @doob1945markoff] and an Event-driven
-implementation.  Both approaches are efficient.  They have similar speed if the 
+implementation.  They have similar speed if the 
 dynamics are Markovian (depending on the network and disease parameters either
 may be faster than the other), but the event-driven implementation can also handle 
 non-Markovian dynamics.  In earlier versions, the event-driven simulations were 
-consistently faster than the Gillespie simulations, and thus they are named 
-`fast_SIR` and `fast_SIS`.  The Gillespie simulations have since been optimized
+consistently faster than the Gillespie simulations, leading to the names
+`fast_SIR` and `fast_SIS`.  The Gillespie simulations have been optimized
 using ideas from [@holme2014model] and [@cota2017optimized].
 
-The algorithms can typically handle an SIR epidemic spreading on 
-hundreds of thousands of individuals in well under a minute on a laptop.  The 
-SIS versions are slower because the number of events that happen is often much
-larger in an SIS simulation.
+The algorithms typically handle an SIR epidemic spreading on 
+hundreds of thousands of individuals in well under a minute on a laptop. 
+SIS versions are slower because the number of events that happen is typically much
+larger.
 
 ### Differential Equations Models
 
-EoN also provides a set of tools to numerically solve approximately 20 differential equations
-models for SIS or SIR disease spread in networks.  The various models use different
-information about the network to make deterministic predictions about
-the number infected at different times.  These use the Scipy integration tools.
[…]
