+++ docs/paper/paper.md
@@ -34,10 +34,10 @@ bibliography: paper.bib
 EoN (EpidemicsOnNetworks) is a pure-python package designed to study
 infectious processes spreading in networks.  It rose out of the 
 book *Mathematics of Epidemics on Networks* [@kiss:EoN], and now consists of over 
-100 user-functions.
+100 user-functions.  EoN relies on the networkx package [@hagberg2008exploring].
 
 
-EoN provides a set of tools for
+EoN provides methods for
 
 - Susceptible-Infected-Susceptible (SIS) and Susceptible-Infected-Recovered 
 (SIR) disease
@@ -52,20 +52,18 @@ EoN provides a set of tools for
 - Stochastic simulation of a wide range of Simple and Complex contagions
 - Visualization and analysis of stochastic simulations
 
-EoN relies on the networkx package [@hagberg2008exploring].
-Its documentation is maintained at 
+The documentation is maintained at 
 https://epidemicsonnetworks.readthedocs.io/en/latest/ 
 including numerous examples at 
-https://epidemicsonnetworks.readthedocs.io/en/latest/Examples.html.  In this 
-paper we provide brief descriptions of a few of EoN's main tools.  The 
-online documentation gives more detail about how to use them, including examples.
+https://epidemicsonnetworks.readthedocs.io/en/latest/Examples.html.  This paper
+briefly describes a few of EoN's main tools.  The online documentation gives 
+full detail, including examples.
 
 ## SIR and SIS disease
 
 ### Stochastic simulation
 The main stochastic simulation tools allow the user to investigate
-standard SIS and SIR dynamics (SEIR/SIRS and other processes are addressed 
-within the simple contagion model):
+standard SIS and SIR dynamics:
 
 - Markovian SIS and SIR simulations  (``fast_SIS``, ``Gillespie_SIS``, ``fast_SIR``, and ``Gillespie_SIR``).
 - non-Markovian SIS and SIR simulations (``fast_nonMarkovian_SIS`` and ``fast_nonMarkovian_SIR``).
@@ -75,15 +73,15 @@ within the simple contagion model):
 For both Markovian and non-Markovian methods it is possible for the transition 
 rates to depend on individual or partnership properties.
 
-The continuous-time stochastic simulations have two different implementations: a 
+The continuous-time stochastic simulations have two implementations: a 
 Gillespie implementation [@gillespie1977exact; @doob1945markoff] and an Event-driven
-implementation.  They have similar speed if the 
-dynamics are Markovian (depending on the network and disease parameters either
-may be faster than the other), but the event-driven implementation can also handle 
+implementation.  They have similar speed if the dynamics are Markovian 
+(depending on network and disease parameters either may be faster than the
+other), but the event-driven implementations also handle 
 non-Markovian dynamics.  In earlier versions, the event-driven simulations were 
-consistently faster than the Gillespie simulations, leading to the names
-`fast_SIR` and `fast_SIS`.  The Gillespie simulations have been optimized
-using ideas from [@holme2014model] and [@cota2017optimized].
+consistently faster, leading to the names `fast_SIR` and `fast_SIS`.  The 
+Gillespie simulations have been optimized using ideas from [@holme2014model] 
+and [@cota2017optimized].
 
 The algorithms typically handle an SIR epidemic spreading on 
 hundreds of thousands of individuals in well under a minute on a laptop. 
@@ -94,13 +92,9 @@ larger.
 
 EoN also provides tools to numerically solve about 20 differential equations
 models for SIS or SIR disease spread in networks using Scipy integration tools.
-The various models use different information about the network to predict the 
-number infected at different times.  Model derivations and explanations of 
+The models use different information about the network to predict the 
+number infected as a function of time.  Model derivations and explanations of 
[…]
