### 2019-12-19 #1731 @joelmiller
@lpantano @arfon @labarba I've dramatically shortened the paper.  Please let me know if there are any further concerns.
