### 2019-12-19 #1731 @labarba
Could you merge some minor copy edits, here? https://github.com/springer-math/Mathematics-of-Epidemics-on-Networks/pull/47
